confirm

+++ Remix-course/app/components/expenses/ExpenseListItem.jsx
@@ -4,10 +4,14 @@ function ExpenseListItem({ id, title, amount }) {
   // const submit = useSubmit();
   const fetcher = useFetcher();
   function deleteExpenseItemHandler() {
+    confirm('Are you sure? Do you want to delete this item?');
     // submit(null,{
     //   method: 'delete',
     //   action: `/expenses/${id}`,
     // });
+    if (!proceed) {
+      return;
+    }
     fetcher.submit(null,
        {method: 'delete',
         action: `/expenses/${id}`,
